LayoutTests/http/tests/security/document-all-expected.txt

CONSOLE MESSAGE: TypeError: Illegal constructor
